btrfs-progs: check: Move more shared codes to mode-common.c
The following functions are mode independent, move them to
mode-common.c.

int zero_log_tree(struct btrfs_root *root);
int reinit_extent_tree(struct btrfs_trans_handle *trans,
                       struct btrfs_fs_info *fs_info);
int btrfs_fsck_reinit_root(struct btrfs_trans_handle *trans,
                           struct btrfs_root *root, int overwrite);
int fill_csum_tree(struct btrfs_trans_handle *trans,
                   struct btrfs_root *csum_root, int search_fs_tree);
int repair_root_items(struct btrfs_fs_info *info);
int do_clear_free_space_cache(struct btrfs_fs_info *fs_info,
                              int clear_version);
bool is_super_size_valid(struct btrfs_fs_info *fs_info);
int check_space_cache(struct btrfs_root *root);
int check_csums(struct btrfs_root *root);
int recow_extent_buffer(struct btrfs_root *root, struct extent_buffer *eb);
